Separate natianlity selector from main player card
