+++ components/profile/player-card/nationality.module.scss
@@ -1,0 +1,15 @@
+.flags {
+  margin: 10px;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  align-content: space-between;
+  gap: 15px;
+  height: 200px;
+  flex-wrap: wrap;
+  overflow-y: scroll;
+}
+
+.flag {
+  cursor: pointer;
+}
